+++ config/transmorpher.php
@@ -36,7 +36,8 @@ return [
 
     'upload' => [
         // Chunk size in mb.
-        'chunk_size' => 1 * 1024 * 1024,
+        // If the server is configured to use S3-Multi-Part uploads, values lower than 5MiB will automatically be set to 5MiB.
+        'chunk_size' => 5 * 1024 * 1024,
         'image' => [
             'validations' => [
                 // Max file size in mb.