+++ src/resources/js/transmorpher.js
@@ -49,7 +49,7 @@ if (!window.transmorpherScriptLoaded) {
             dictFileTooBig: medium.translations['max_file_size_exceeded'],
             dictInvalidFileType: medium.translations['invalid_file_type'],
             createImageThumbnails: false,
-            ...handler.getDropzoneOptions(),
+            ...handler.getDropzoneOptions({transmorpherMedium: medium}),
             init: function () {
                 // Processing-Event is emitted when the upload starts.
                 this.on('processing', function () {